FHAC-622: Refactoring of DocSubmissionAuditTransforms class
